consolidate chat message list
