+++ src/stores/ui.js
@@ -250,6 +250,9 @@ isTerminalMaximizedStore.listen((value) => {
   }
 });
 
+// State to track when the chat is processing a prompt
+export const isChatProcessing = atom(false);
+
 // Image overlay state
 export const imageOverlayStore = atom({
   isOpen: false,
